Extract Tabs engine into TabManager (pure refactor, no behavior change)
Cut #1 of the FrameManager decomposition roadmap (discussion #129).
Moves the Tabs engine (~1,315 lines) out of the 11,189-line Framemanager
god class into a new static class TabManager:

  RefreshTabStripUI, AddNewTab, RenameTab, DeleteTab, SwitchTabByFrame,
  MoveTab, SynchronizeTab0Content, AreItemArraysEqual, RefreshTabStyling,
  RefreshTabColors, ApplyTabStyle, Tab_MouseEnter/Leave_Lambda,
  plus the _isSynchronizing field.

Pure move + call-site qualifier changes only; no method bodies modified.
The single non-move change is SynchronizeTab0Content private -> public,
required because CreateFrame (still in Framemanager) calls it.

Re-applies PR #130 on the new 2.7.8.352 baseline.
FrameManager.cs: 11,189 -> 9,963 lines.
